flask_backend/requirements.txt
Flask
flask-cors
tensorflow==2.15.0
pillow
